[1.18.2] Migration of the mod to the Minecraft 1.18.2 version

+++ src/main/java/power/keepeersofthestones/procedures/TimeCheckpointProcedure.java
@@ -16,21 +16,21 @@ public class TimeCheckpointProcedure {
 			_serverPlayer.setRespawnPosition(_serverPlayer.level.dimension(), new BlockPos((int) x, (int) (y + 1), (int) z), _serverPlayer.getYRot(),
 					true, false);
 		{
-			double _setval = (double) (y + 1);
+			double _setval = y + 1;
 			entity.getCapability(PowerModVariables.PLAYER_VARIABLES_CAPABILITY, null).ifPresent(capability -> {
 				capability.spawny = _setval;
 				capability.syncPlayerVariables(entity);
 			});
 		}
 		{
-			double _setval = (double) x;
+			double _setval = x;
 			entity.getCapability(PowerModVariables.PLAYER_VARIABLES_CAPABILITY, null).ifPresent(capability -> {
 				capability.spawnx = _setval;
 				capability.syncPlayerVariables(entity);
 			});
 		}
 		{
-			double _setval = (double) z;
+			double _setval = z;
 			entity.getCapability(PowerModVariables.PLAYER_VARIABLES_CAPABILITY, null).ifPresent(capability -> {
 				capability.spawnz = _setval;
 				capability.syncPlayerVariables(entity);
